Have made a notes service which handles fetching, adding and deleting notes. Have modified the code so async pipe isnt used anywhere. Instead subsciptions are used. Cleaned up globals service so all notes related stuff is in notes service and all authentication related stuff is in auth service.
Next:
Add method in notes.service.ts for updating existing notes.
