+++ src/app/services/notes.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { NotesService } from './notes.service';
+
+describe('NotesService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [NotesService]
+    });
+  });
+
+  it('should be created', inject([NotesService], (service: NotesService) => {
+    expect(service).toBeTruthy();
+  }));
+});
